Module Tezos_client_002_PsYLVpVv.Michelson_v1_printer

val print_expr : Stdlib.Format.formatter -> Tezos_raw_protocol_002_PsYLVpVv.Script_repr.expr -> unit

val print_expr_unwrapped : Stdlib.Format.formatter -> Tezos_raw_protocol_002_PsYLVpVv.Script_repr.expr -> unit

val print_execution_trace : Stdlib.Format.formatter -> (Tezos_raw_protocol_002_PsYLVpVv.Alpha_context.Script.location * Tezos_raw_protocol_002_PsYLVpVv.Alpha_context.Gas.t * (Tezos_raw_protocol_002_PsYLVpVv.Alpha_context.Script.expr * string option) list) list -> unit

val inject_types : Tezos_raw_protocol_002_PsYLVpVv.Script_tc_errors.type_map -> Michelson_v1_parser.parsed -> Tezos_micheline.Micheline_printer.node

Insert the type map returned by the typechecker as comments in a printable Micheline AST.

val unparse_toplevel : ?⁠type_map:Tezos_raw_protocol_002_PsYLVpVv.Script_tc_errors.type_map -> Tezos_raw_protocol_002_PsYLVpVv.Alpha_context.Script.expr -> Michelson_v1_parser.parsed

Unexpand the macros and produce the result of parsing an intermediate pretty printed source. Useful when working with contracts extracted from the blockchain and not local files.

val unparse_expression : Tezos_raw_protocol_002_PsYLVpVv.Alpha_context.Script.expr -> Michelson_v1_parser.parsed

val unparse_invalid : string Tezos_micheline.Micheline.canonical -> Michelson_v1_parser.parsed

Unexpand the macros and produce the result of parsing an intermediate pretty printed source. Works on generic trees,for programs that fail to be converted to a specific script version.